refactor resize method and unit test
